src/main/webapp/app/entities/parametrage-global/delete/parametrage-global-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ParametrageGlobalService } from '../service/parametrage-global.service';

import { ParametrageGlobalDeleteDialogComponent } from './parametrage-global-delete-dialog.component';

describe('ParametrageGlobal Management Delete Component', () => {
  let comp: ParametrageGlobalDeleteDialogComponent;
  let fixture: ComponentFixture<ParametrageGlobalDeleteDialogComponent>;
  let service: ParametrageGlobalService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [ParametrageGlobalDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(ParametrageGlobalDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ParametrageGlobalDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ParametrageGlobalService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
